renamed apicontroller to logcontroller and moved its directory
