+++ routes/web.php
@@ -152,5 +152,5 @@ Route::group(['middleware' => ['permission:Can view admin page']], function(){
 
 // Apply the API middleware
 Route::group(['middleware' => ['application_api']], function(){
-	Route::get('/api/log', 'ApiController@getLog');
+	Route::get('/api/log', 'Api\LogController@getLog');
 });
